refactor(stats): split StatsComponents.kt into four focused files
StatsComponents.kt (942 lines) → split into:
- StatsSharedUi.kt       — SectionHeader, StatHeadlineCard, MultipleLinearProgressIndicator, NeverWornSection, empty states
- StatsActivityComponents.kt  — period selector, headline cards, most worn, cost per wear, total logs, Vico chart
- StatsCompositionComponents.kt — wash status, category/subcategory/occasion/color breakdowns, BreakdownSection helper
- SegmentedBar.kt        — BarSegment, TooltipContent, withOtherGroup, SegmentedBar, SegmentTooltip

No behaviour changes — pure file reorganisation.
